Changes to the Application process
The first commit in a series of changes to the Application process. This one opens it up to the public, adds the user meta fields to fill out always (even if they're logged in), and puts the entire process in hide/show stepped divs as per the received PDF.

It's broken though. I haven't changed any backend code, so the user meta doesn't get saved, and there are bugs that existed prior to my work. Still in progress.